Moving Bigtable helpers for duration protobufs into core.
