+++ bigtable/google/cloud/bigtable/column_family.py
@@ -15,60 +15,11 @@
 """User friendly container for Google Cloud Bigtable Column Family."""
 
 
-import datetime
-
-from google.protobuf import duration_pb2
-
+from google.cloud import _helpers
 from google.cloud.bigtable._generated import (
     table_pb2 as table_v2_pb2)
 from google.cloud.bigtable._generated import (
     bigtable_table_admin_pb2 as table_admin_v2_pb2)
-
-
-def _timedelta_to_duration_pb(timedelta_val):
-    """Convert a Python timedelta object to a duration protobuf.
-
-    .. note::
-
-        The Python timedelta has a granularity of microseconds while
-        the protobuf duration type has a duration of nanoseconds.
-
-    :type timedelta_val: :class:`datetime.timedelta`
-    :param timedelta_val: A timedelta object.
-
-    :rtype: :class:`google.protobuf.duration_pb2.Duration`
-    :returns: A duration object equivalent to the time delta.
-    """
-    seconds_decimal = timedelta_val.total_seconds()
-    # Truncate the parts other than the integer.
-    seconds = int(seconds_decimal)
-    if seconds_decimal < 0:
-        signed_micros = timedelta_val.microseconds - 10**6
-    else:
-        signed_micros = timedelta_val.microseconds
-    # Convert nanoseconds to microseconds.
-    nanos = 1000 * signed_micros
-    return duration_pb2.Duration(seconds=seconds, nanos=nanos)
-
-
-def _duration_pb_to_timedelta(duration_pb):
-    """Convert a duration protobuf to a Python timedelta object.
-
-    .. note::
-
-        The Python timedelta has a granularity of microseconds while
-        the protobuf duration type has a duration of nanoseconds.
-
-    :type duration_pb: :class:`google.protobuf.duration_pb2.Duration`
-    :param duration_pb: A protobuf duration object.
-
-    :rtype: :class:`datetime.timedelta`
-    :returns: The converted timedelta object.
-    """
-    return datetime.timedelta(
-        seconds=duration_pb.seconds,
-        microseconds=(duration_pb.nanos / 1000.0),
-    )
 
 
 class GarbageCollectionRule(object):
@@ -137,7 +88,7 @@ class MaxAgeGCRule(GarbageCollectionRule):
         :rtype: :class:`.table_v2_pb2.GcRule`
         :returns: The converted current object.
         """
-        max_age = _timedelta_to_duration_pb(self.max_age)
+        max_age = _helpers._timedelta_to_duration_pb(self.max_age)
         return table_v2_pb2.GcRule(max_age=max_age)
 
 
@@ -325,7 +276,7 @@ def _gc_rule_from_pb(gc_rule_pb):
     if rule_name == 'max_num_versions':
         return MaxVersionsGCRule(gc_rule_pb.max_num_versions)
     elif rule_name == 'max_age':
-        max_age = _duration_pb_to_timedelta(gc_rule_pb.max_age)
+        max_age = _helpers._duration_pb_to_timedelta(gc_rule_pb.max_age)
         return MaxAgeGCRule(max_age)
     elif rule_name == 'union':
         return GCRuleUnion([_gc_rule_from_pb(rule)
