+++ bigtable/unit_tests/test_column_family.py
@@ -16,75 +16,6 @@
 import unittest
 
 
-class Test__timedelta_to_duration_pb(unittest.TestCase):
-
-    def _call_fut(self, *args, **kwargs):
-        from google.cloud.bigtable.column_family import (
-            _timedelta_to_duration_pb)
-
-        return _timedelta_to_duration_pb(*args, **kwargs)
-
-    def test_it(self):
-        import datetime
-        from google.protobuf import duration_pb2
-
-        seconds = microseconds = 1
-        timedelta_val = datetime.timedelta(seconds=seconds,
-                                           microseconds=microseconds)
-        result = self._call_fut(timedelta_val)
-        self.assertIsInstance(result, duration_pb2.Duration)
-        self.assertEqual(result.seconds, seconds)
-        self.assertEqual(result.nanos, 1000 * microseconds)
-
-    def test_with_negative_microseconds(self):
-        import datetime
-        from google.protobuf import duration_pb2
-
-        seconds = 1
-        microseconds = -5
-        timedelta_val = datetime.timedelta(seconds=seconds,
-                                           microseconds=microseconds)
-        result = self._call_fut(timedelta_val)
-        self.assertIsInstance(result, duration_pb2.Duration)
-        self.assertEqual(result.seconds, seconds - 1)
-        self.assertEqual(result.nanos, 10**9 + 1000 * microseconds)
-
-    def test_with_negative_seconds(self):
-        import datetime
-        from google.protobuf import duration_pb2
-
-        seconds = -1
-        microseconds = 5
-        timedelta_val = datetime.timedelta(seconds=seconds,
-                                           microseconds=microseconds)
-        result = self._call_fut(timedelta_val)
-        self.assertIsInstance(result, duration_pb2.Duration)
-        self.assertEqual(result.seconds, seconds + 1)
-        self.assertEqual(result.nanos, -(10**9 - 1000 * microseconds))
-
-
-class Test__duration_pb_to_timedelta(unittest.TestCase):
-
-    def _call_fut(self, *args, **kwargs):
-        from google.cloud.bigtable.column_family import (
-            _duration_pb_to_timedelta)
-
-        return _duration_pb_to_timedelta(*args, **kwargs)
-
-    def test_it(self):
-        import datetime
-        from google.protobuf import duration_pb2
-
-        seconds = microseconds = 1
-        duration_pb = duration_pb2.Duration(seconds=seconds,
-                                            nanos=1000 * microseconds)
-        timedelta_val = datetime.timedelta(seconds=seconds,
-                                           microseconds=microseconds)
-        result = self._call_fut(duration_pb)
-        self.assertIsInstance(result, datetime.timedelta)
-        self.assertEqual(result, timedelta_val)
-
-
 class TestMaxVersionsGCRule(unittest.TestCase):
 
     @staticmethod
